chore: migrate `addStyleTag`
