+++ docs/api/puppeteer.frame.addstyletag_1.md
@@ -1,0 +1,25 @@
+---
+sidebar_label: Frame.addStyleTag_1
+---
+
+# Frame.addStyleTag() method
+
+**Signature:**
+
+```typescript
+class Frame {
+  addStyleTag(
+    options: FrameAddStyleTagOptions
+  ): Promise<ElementHandle<HTMLLinkElement>>;
+}
+```
+
+## Parameters
+
+| Parameter | Type                                                              | Description |
+| --------- | ----------------------------------------------------------------- | ----------- |
+| options   | [FrameAddStyleTagOptions](./puppeteer.frameaddstyletagoptions.md) |             |
+
+**Returns:**
+
+Promise&lt;[ElementHandle](./puppeteer.elementhandle.md)&lt;HTMLLinkElement&gt;&gt;
