+++ docs/api/puppeteer.page.addstyletag_1.md
@@ -1,0 +1,25 @@
+---
+sidebar_label: Page.addStyleTag_1
+---
+
+# Page.addStyleTag() method
+
+**Signature:**
+
+```typescript
+class Page {
+  addStyleTag(
+    options: FrameAddStyleTagOptions
+  ): Promise<ElementHandle<HTMLLinkElement>>;
+}
+```
+
+## Parameters
+
+| Parameter | Type                                                              | Description |
+| --------- | ----------------------------------------------------------------- | ----------- |
+| options   | [FrameAddStyleTagOptions](./puppeteer.frameaddstyletagoptions.md) |             |
+
+**Returns:**
+
+Promise&lt;[ElementHandle](./puppeteer.elementhandle.md)&lt;HTMLLinkElement&gt;&gt;
