+++ docs/api/puppeteer.page.md
@@ -82,7 +82,8 @@ page.off('request', logRequest);
 | [$eval(selector, pageFunction, args)](./puppeteer.page._eval.md)                            |           | This method runs <code>document.querySelector</code> within the page and passes the result as the first argument to the <code>pageFunction</code>.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 |
 | [$x(expression)](./puppeteer.page._x.md)                                                    |           | The method evaluates the XPath expression relative to the page document as its context node. If there are no such elements, the method resolves to an empty array.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 |
 | [addScriptTag(options)](./puppeteer.page.addscripttag.md)                                   |           | Adds a <code>&lt;script&gt;</code> tag into the page with the desired URL or content.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              |
-| [addStyleTag(options)](./puppeteer.page.addstyletag.md)                                     |           | Adds a <code>&lt;link rel=&quot;stylesheet&quot;&gt;</code> tag into the page with the desired URL or a <code>&lt;style type=&quot;text/css&quot;&gt;</code> tag with the content.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 |
+| [addStyleTag(options)](./puppeteer.page.addstyletag.md)                                     |           | <p>Adds a <code>&lt;link rel=&quot;stylesheet&quot;&gt;</code> tag into the page with the desired URL or a <code>&lt;style type=&quot;text/css&quot;&gt;</code> tag with the content.</p><p>Shortcut for .</p>                                                                                                                                                                                                                                                                                                                                                                                                                                                     |
+| [addStyleTag(options)](./puppeteer.page.addstyletag_1.md)                                   |           |                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    |
 | [authenticate(credentials)](./puppeteer.page.authenticate.md)                               |           | Provide credentials for <code>HTTP authentication</code>.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          |
 | [bringToFront()](./puppeteer.page.bringtofront.md)                                          |           | Brings page to front (activates tab).                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              |
 | [browser()](./puppeteer.page.browser.md)                                                    |           | Get the browser the page belongs to.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               |
